chore: put insert and search in different files
